TDD 적용, lombok 최적화 및 기타 정리..
생성자 주입 방식으로 변경 @Autowired -> @RequiredArgsConstructor
Integer id 자료형 변경  DTO -> long, entity -> Long = 가독성(null 가능), 범위
TDD 작성 및 test 설정 추가 (h2)

+++ src/main/java/io/ybg/demo/repository/MemberRepo.java
@@ -3,6 +3,6 @@ package io.ybg.demo.repository;
 import io.ybg.demo.entity.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepo extends JpaRepository<MemberEntity, Integer> {
+public interface MemberRepo extends JpaRepository<MemberEntity, Long> {
     boolean existsByEmail(String email);
 }
